fix: ./sasjs only has doxy folder
